Add files via upload
Initial upload of the down-arrow visualizer.  Right now the user can view the graph ideal for any networkX defined graph by running the cells in the .ipynb file.  The user can click on a vertex of the ideal to see the associated graph displayed to the right.  All edges of the ideal are directed and point upwards.  They indicate subgraph inclusion of associated graphs moving from lower vertex to upper vertex.